Code cleanup and documentation.
